+++ Library/OcAppleEventLib/PointerHandler.c
@@ -47,11 +47,16 @@ WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 #define MIN_POINTER_POLL_PERIOD  10
 #define MAX_POINTER_POLL_PERIOD  80
 
+#define MAX_CLICK_DURATION      748  // 374 for 2 ms
+#define MAX_DOUBLE_CLICK_SPEED  148  // 74 for 2 ms
+
+#define MAX_POLL_DURATION  ((MAX_UINT32 / 10000) / MAX_DOUBLE_CLICK_SPEED)
+
 GLOBAL_REMOVE_IF_UNREFERENCED UINT32  mPointerSpeedDiv = 0;
 GLOBAL_REMOVE_IF_UNREFERENCED UINT32  mPointerSpeedMul = 0;
 
-STATIC UINT16  mMaximumDoubleClickSpeed = 75; // 374 for 2 ms
-STATIC UINT16  mMaximumClickDuration    = 15; // 74 for 2 ms
+STATIC UINT16  mMaximumDoubleClickSpeed = 0;
+STATIC UINT16  mMaximumClickDuration    = 0;
 
 // MINIMAL_MOVEMENT
 #define MINIMAL_MOVEMENT  5
@@ -91,9 +96,9 @@ STATIC UINTN  mNumberOfPointerProtocols = 0;
 STATIC EFI_EVENT  mSimplePointerPollEvent = NULL;
 
 // mSimplePointerPollTime
-STATIC UINT64  mSimplePointerPollTime;
-STATIC UINT64  mSimplePointerMinPollTime = MIN_POINTER_POLL_PERIOD * 10000;
-STATIC UINT64  mSimplePointerMaxPollTime = MAX_POINTER_POLL_PERIOD * 10000;
+STATIC UINT32  mSimplePointerPollTime    = 0;
+STATIC UINT32  mSimplePointerMinPollTime = 0;
+STATIC UINT32  mSimplePointerMaxPollTime = 0;
 
 STATIC UINT32  mSimplePointerPollMask = POINTER_POLL_ALL_MASK;
 
@@ -136,8 +141,8 @@ STATIC DIMENSION  mResolution = { 800, 600 };
 STATIC UINT64  mMaxPointerResolutionX = 1;
 STATIC UINT64  mMaxPointerResolutionY = 1;
 
-STATIC INT64  mPointerRawX;
-STATIC INT64  mPointerRawY;
+STATIC UINT64  mPointerRawX;
+STATIC UINT64  mPointerRawY;
 
 VOID
 InternalSetPointerPolling (
@@ -148,15 +153,22 @@ InternalSetPointerPolling (
 {
   if (PointerPollMin == POINTER_POLL_DEFAULT) {
     PointerPollMin = MIN_POINTER_POLL_PERIOD;
+  } else if (PointerPollMin > MAX_POLL_DURATION) {
+    PointerPollMin = MAX_POLL_DURATION;
   }
 
-  mSimplePointerMinPollTime = EFI_TIMER_PERIOD_MILLISECONDS (PointerPollMin);
+  mSimplePointerMinPollTime = PointerPollMin * 10000;
+
+  mMaximumClickDuration    = (UINT16)(MAX_CLICK_DURATION / PointerPollMin);
+  mMaximumDoubleClickSpeed = (UINT16)(MAX_DOUBLE_CLICK_SPEED / PointerPollMin);
 
   if (PointerPollMax == POINTER_POLL_DEFAULT) {
     PointerPollMax = MAX_POINTER_POLL_PERIOD;
+  } else if (PointerPollMax > MAX_POLL_DURATION) {
+    PointerPollMax = MAX_POLL_DURATION;
   }
 
-  mSimplePointerMaxPollTime = EFI_TIMER_PERIOD_MILLISECONDS (PointerPollMax);
+  mSimplePointerMaxPollTime = PointerPollMax * 10000;
 
   mSimplePointerPollMask = PointerPollMask;
 }
@@ -225,20 +237,20 @@ InternalRegisterSimplePointerInterface (
     mPointerProtocols = Instance;
 
     if (SimplePointer->Mode->ResolutionX > mMaxPointerResolutionX) {
-      mPointerRawX = MultS64x64 (mPointerRawX, (INT64)mMaxPointerResolutionX);
-      mPointerRawX = DivS64x64Remainder (
+      mPointerRawX = MultU64x64 (mPointerRawX, mMaxPointerResolutionX);
+      mPointerRawX = DivU64x64Remainder (
                        mPointerRawX,
-                       (INT64)SimplePointer->Mode->ResolutionX,
+                       SimplePointer->Mode->ResolutionX,
                        NULL
                        );
       mMaxPointerResolutionX = SimplePointer->Mode->ResolutionX;
     }
 
     if (SimplePointer->Mode->ResolutionY > mMaxPointerResolutionY) {
-      mPointerRawY = MultS64x64 (mPointerRawY, (INT64)mMaxPointerResolutionY);
-      mPointerRawY = DivS64x64Remainder (
+      mPointerRawY = MultU64x64 (mPointerRawY, mMaxPointerResolutionY);
+      mPointerRawY = DivU64x64Remainder (
                        mPointerRawY,
-                       (INT64)SimplePointer->Mode->ResolutionY,
+                       SimplePointer->Mode->ResolutionY,
                        NULL
                        );
       mMaxPointerResolutionY = SimplePointer->Mode->ResolutionY;
@@ -735,8 +747,9 @@ InternalSimplePointerPollNotifyFunction (
   UINT64                       EndTime;
   INT64                        MaxRawPointerX;
   INT64                        MaxRawPointerY;
-  UINT64                       ClickTemp;
-  UINT64                       DoubleClickTemp;
+  UINT32                       ClickTemp;
+  UINT32                       DoubleClickTemp;
+  UINT64                       SimplePointerPollTime;
 
   StartTime = GetPerformanceCounter ();
 
@@ -803,36 +816,40 @@ InternalSimplePointerPollNotifyFunction (
         //
         // CHANGE: Fix maximum coordinates.
         //
-        mPointerRawX  += UiScaleX;
-        MaxRawPointerX = MultS64x64 (
+        ScaledX        = mPointerRawX + UiScaleX;
+        MaxRawPointerX = MultU64x64 (
                            mResolution.Horizontal - 1,
-                           (INT64)mMaxPointerResolutionX
+                           mMaxPointerResolutionX
                            );
-        if (mPointerRawX > MaxRawPointerX) {
-          mPointerRawX = MaxRawPointerX;
-        } else if (mPointerRawX < 0) {
-          mPointerRawX = 0;
+        if (ScaledX > MaxRawPointerX) {
+          ScaledX = MaxRawPointerX;
+        } else if (ScaledX < 0) {
+          ScaledX = 0;
         }
 
-        mPointerRawY  += UiScaleY;
+        mPointerRawX = (UINT64)ScaledX;
+
+        ScaledY        = mPointerRawY + UiScaleY;
         MaxRawPointerY = MultS64x64 (
                            mResolution.Vertical - 1,
                            (INT64)mMaxPointerResolutionY
                            );
-        if (mPointerRawY > MaxRawPointerY) {
-          mPointerRawY = MaxRawPointerY;
-        } else if (mPointerRawY < 0) {
-          mPointerRawY = 0;
+        if (ScaledY > MaxRawPointerY) {
+          ScaledY = MaxRawPointerY;
+        } else if (ScaledY < 0) {
+          ScaledY = 0;
         }
 
-        ScaledX = DivS64x64Remainder (
-                    mPointerRawX,
-                    (INT64)mMaxPointerResolutionX,
+        mPointerRawY = (UINT64)ScaledY;
+
+        ScaledX = DivU64x64Remainder (
+                    ScaledX,
+                    mMaxPointerResolutionX,
                     NULL
                     );
 
         ScaledY = DivS64x64Remainder (
-                    mPointerRawY,
+                    ScaledY,
                     (INT64)mMaxPointerResolutionY,
                     NULL
                     );
@@ -901,28 +918,19 @@ InternalSimplePointerPollNotifyFunction (
 
     EndTime = GetTimeInNanoSecond (EndTime - StartTime);
     // Maximum time allowed in this function is half the interval plus some margin (0.55 * 100ns)
-    if (EndTime > mSimplePointerPollTime * 55ULL) {
-      ClickTemp       = MultU64x32 (mSimplePointerPollTime, mMaximumClickDuration);
-      DoubleClickTemp = MultU64x32 (
-                          mSimplePointerPollTime,
-                          mMaximumDoubleClickSpeed
-                          );
+    if (EndTime > mSimplePointerPollTime * 55) {
+      ClickTemp       = mSimplePointerPollTime * mMaximumClickDuration;
+      DoubleClickTemp = mSimplePointerPollTime * mMaximumDoubleClickSpeed;
 
-      mSimplePointerPollTime = DivU64x32 (EndTime, 50);
-      if (mSimplePointerPollTime > mSimplePointerMaxPollTime) {
+      SimplePointerPollTime = DivU64x32 (EndTime, 50);
+      if (SimplePointerPollTime <= mSimplePointerMaxPollTime) {
+        mSimplePointerPollTime = (UINT32)SimplePointerPollTime;
+      } else {
         mSimplePointerPollTime = mSimplePointerMaxPollTime;
       }
 
-      mMaximumClickDuration = (UINT16)DivU64x64Remainder (
-                                        ClickTemp,
-                                        mSimplePointerPollTime,
-                                        NULL
-                                        );
-      mMaximumDoubleClickSpeed = (UINT16)DivU64x64Remainder (
-                                           DoubleClickTemp,
-                                           mSimplePointerPollTime,
-                                           NULL
-                                           );
+      mMaximumClickDuration    = (UINT16)(ClickTemp / mSimplePointerPollTime);
+      mMaximumDoubleClickSpeed = (UINT16)(DoubleClickTemp / mSimplePointerPollTime);
 
       gBS->SetTimer (mSimplePointerPollEvent, TimerPeriodic, mSimplePointerPollTime);
     }